kernel signer class
