+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@dimo-network/dimo-node-sdk",
-  "version": "1.2.2",
+  "version": "1.1.9",
   "description": "DIMO SDK",
   "keywords": [
     "DIMO"
