+++ src/index.ts
@@ -5,12 +5,11 @@ export type {
   ENVIRONMENT,
 } from "./core/types/interface.js";
 export {
-  mintVehicleCallData,
   mintVehicleWithDeviceDefinition,
-  setPermissionsSACD,
-  setVehiclePermissions,
-  sendDIMOTokensCallData,
-  sendDIMOTokens,
-} from "./DIMODeveloperSDK.js";
+  mintVehicleCallData,
+} from "./core/actions/mintVehicleWithDeviceDefinition.js";
+export { setVehiclePermissions, setPermissionsSACD } from "./core/actions/setPermissionsSACD.js";
+export { sendDIMOTokens, sendDIMOTokensCallData } from "./core/actions/sendDIMOTokens.js";
 export { kernelClientFromPasskeySigner } from "./core/kernelClientFromSigner/kernelClientFromPasskey.js";
 export { kernelClientFromPrivateKey } from "./core/kernelClientFromSigner/kernelClientFromPrivateKey.js";
+export { KernelSigner } from "./DIMODeveloperSDK.js";
