+++ tsconfig.base.json
@@ -27,6 +27,8 @@
     "lib": [
       "esnext",
       "ES2016",
+      "dom",
+      "dom.iterable"
     ],
     "module": "NodeNext",
     "moduleResolution": "NodeNext",
